0. Spring Boot [#348573]

+++ src/main/java/ru/job4j/forum/repositories/PostRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.job4j.forum.enity.Post;
 
 @Repository
-public interface PostRepository extends CrudRepository<Post, Long> {
+public interface PostRepository extends CrudRepository<Post, Integer> {
 }
